+++ src/Components/Atoms/Link/Link.tsx
@@ -8,8 +8,6 @@ import styles from './Link.module.scss';
 export interface ILinkProps extends NavLinkProps {
   /** External link indicator (optional, default: false) */
   externalLink?: boolean;
-  /** Text displayed for the link (optional, default: undefined) */
-  text?: string;
   /** Url of the destination in a string or Location format */
   to: To;
   /** Tooltip content of the link (optional, default: undefined) */
@@ -17,7 +15,7 @@ export interface ILinkProps extends NavLinkProps {
 }
 
 const Link = (props: PropsWithChildren<ILinkProps>): ReactElement => {
-  const { children, externalLink, to, text, tooltip, ...rest } = props;
+  const { children, externalLink, to, tooltip, ...rest } = props;
 
   return (
     <div className={styles.container}>
@@ -27,10 +25,7 @@ const Link = (props: PropsWithChildren<ILinkProps>): ReactElement => {
         rel={externalLink ? 'noreferrer' : rest.rel}
         target={externalLink ? '_blank' : rest.target}>
         <Tooltip tooltip={tooltip}>
-          <>
-            {text}
-            {children}
-          </>
+          <>{children}</>
         </Tooltip>
         {externalLink && <Icon icon={['fal', 'external-link']} className={styles.icon} size='1x' />}
       </NavLink>
@@ -40,7 +35,6 @@ const Link = (props: PropsWithChildren<ILinkProps>): ReactElement => {
 
 Link.defaultProps = {
   externalLink: false,
-  text: undefined,
   tooltip: undefined,
 };