+++ src/Components/Atoms/Link/__tests__/Link.test.tsx
@@ -5,12 +5,16 @@ import Link from '../Link';
 
 describe('Link Component', () => {
   it('Link renders internal link', () => {
-    const { container } = render(<Link text='TEXT' to='/link' />);
+    const { container } = render(<Link to='/link'>TEXT</Link>);
     expect(container).toMatchSnapshot();
   });
 
   it('Link renders external link with tooltip', () => {
-    const { container } = render(<Link text='TEXT' to='http://www.link.com' tooltip='TOOLTIP' externalLink />);
+    const { container } = render(
+      <Link to='http://www.link.com' tooltip='TOOLTIP' externalLink>
+        TEXT
+      </Link>,
+    );
     expect(container).toMatchSnapshot();
   });
 });